+++ CHANGELOG.md
@@ -12,5 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Invoke-SubDirectoryGitCommand function (#1).
 - Add Write-GitRepositoryDetail function (#1).
 - Add Remove-BuildArtifact function (#2).
+- Add Resolve-Error function (#3).
+- Add Stop-Wsl function (#3).
 
-[unreleased]: https://github.com/lancra/pwsh/compare/v0.1.0-preview-0...HEAD
+[unreleased]: https://github.com/lancra/pwsh/compare/v0.1.0-preview0...HEAD